fix: align runner app name, userData path, and decouple net-helpers from vitest
- spec/_vitest_runner/package.json: name/productName match spec/package.json
  so app.name assertions hold
- worker-entry: userData is <mkdtemp>/<app.name>/ so getPath('userData')
  still includes the app name
- net-helpers: import defer from defer-helpers and own listen() directly,
  so the utility-process fixture (which ts-node-requires net-helpers) no
  longer pulls in spec-helpers -> vitest
